feat: 更新 wxbot-ewbhook 至 v2.6.0
