+++ config.yaml.example
@@ -9,6 +9,7 @@ wechatter_port: 4000
 # WX Webhook
 wx_webhook_base_api: http://localhost:3001
 wx_webhook_recv_api_path: /receive_msg
+wx_webhook_token: "your_wx_webhook_token"
 
 
 # Admin
